requirements.txt
playwright>=1.58
playwright-stealth>=2.0
tzdata>=2026.2
requests>=2.32
python-dotenv>=1.0
anthropic>=0.50
python-telegram-bot>=22.0
